refactor jwt mod to avoid in-file mod.
